Backend/requirements.txt
flask
pillow
piexif
flask-cors